requirements.full.txt
asgiref==3.5.2
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2

# unneccessary dependencies
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
black==22.6.0
click==8.1.3
tomli==2.0.1
psycopg2-binary==2.9.3